Refactor hive-outer-layout out
